+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Rendering room upgrades (tombstone event)
+- API: add multi-profile option to `MultiMessengerConfiguration`
 
 ### Changed